+++ tuic-client/src/main.rs
@@ -1,14 +1,16 @@
-use socks5::Server;
-
 use self::{
     config::{Config, ConfigError},
     connection::Endpoint,
+    socks5::Server,
 };
-use std::{env, process};
+use quinn::{ConnectError, ConnectionError};
+use std::{env, io::Error as IoError, process};
+use thiserror::Error;
+use tuic_quinn::Error as ModelError;
+use webpki::Error as WebpkiError;
 
 mod config;
 mod connection;
-mod error;
 mod socks5;
 mod utils;
 
@@ -43,4 +45,26 @@ async fn main() {
     }
 
     Server::start().await;
+}
+
+#[derive(Debug, Error)]
+pub enum Error {
+    #[error(transparent)]
+    Io(#[from] IoError),
+    #[error(transparent)]
+    Connect(#[from] ConnectError),
+    #[error(transparent)]
+    Connection(#[from] ConnectionError),
+    #[error(transparent)]
+    Model(#[from] ModelError),
+    #[error(transparent)]
+    Webpki(#[from] WebpkiError),
+    #[error("timeout establishing connection")]
+    Timeout,
+    #[error("cannot resolve the server name")]
+    DnsResolve,
+    #[error("received packet from an unexpected source")]
+    WrongPacketSource,
+    #[error("invalid socks5 authentication")]
+    InvalidSocks5Auth,
 }
